Small fixes. Match 2020 plugin changes to 2021
